backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
transformers==4.35.0
torch==2.1.0
opencv-python==4.8.1.78
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
